requirements.txt
pyrogram
tgcrypto
sqlalchemy
asyncpg
aiohttp
sqlalchemy-utils
pyyaml
openai
python-dotenv
